refactor(swebench): separate Pyxis worker
